requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
python-dotenv